Run the chorus and delay in double too
Same divergence as the reverb tank, in the other two send effects.  Upstream
holds its chorus and delay state in JavaScript numbers and only narrows when a
sample lands in a Float32Array delay buffer; the port held the state itself in
floats and rounded at every step of a feedback path.

Widen both: the chorus's LFO phase and rate, its pre-lowpass state, depth,
delay and feedback, the read positions and their interpolation, and the
mixdown; and the delay's pre-lowpass, its per-line feedback and gain, and the
sample it circulates.  Buffers stay float32 on both sides and the writes narrow
on store, so the port keeps exactly what upstream keeps.

The chorus axis was 80 dB and now reads 94.0 flat across send level.  On the
corpus, gs_chorus_level goes from 25.35 dB to 60.41 and gs_chorus_rate from
19.21 to 23.50, with nothing else moving.  The delay axis was already at the
level the reverb fix lifted everything to, so its own change is not separately
visible here, but it had the same defect and would have shown on a probe that
drove it harder.

Every grid axis outside the two repeated-note ones now sits at 92 dB or better.
Those two are upstream's voice recycling inheriting state across notes, which
is not ours to fix -- see the patch drafted for upstream.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/delay.h
@@ -33,9 +33,9 @@ typedef struct SS_DelayParams {
 typedef struct SS_Delay {
 	SS_DelayParams parameters;
 
-	float preLPFfc;
-	float preLPFa;
-	float preLPFz;
+	double preLPFfc;
+	double preLPFa;
+	double preLPFz;
 
 	SS_DelayLine *delayLeft;
 	SS_DelayLine *delayRight;
@@ -47,12 +47,12 @@ typedef struct SS_Delay {
 	float *delay_center_output;
 	float *delay_pre_lpf;
 
-	float delay_center_time;
-	float delay_left_multiplier;
-	float delay_right_multiplier;
+	double delay_center_time;
+	double delay_left_multiplier;
+	double delay_right_multiplier;
 
-	float gain;
-	float reverb_gain;
+	double gain;
+	double reverb_gain;
 } SS_Delay;
 
 SS_Delay SPESSASYNTH_EXPORTS *ss_delay_create(float sample_rate, int max_buffer_size);
